+++ api/bootstrap/kubeadm/v1beta1/conversion.go
@@ -17,276 +17,16 @@ limitations under the License.
 package v1beta1
 
 import (
-	"fmt"
 	"reflect"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	apimachineryconversion "k8s.io/apimachinery/pkg/conversion"
 	"k8s.io/utils/ptr"
-	"sigs.k8s.io/controller-runtime/pkg/conversion"
 
 	bootstrapv1 "sigs.k8s.io/cluster-api/api/bootstrap/kubeadm/v1beta2"
 	clusterv1beta1 "sigs.k8s.io/cluster-api/api/core/v1beta1"
 	clusterv1 "sigs.k8s.io/cluster-api/api/core/v1beta2"
-	utilconversion "sigs.k8s.io/cluster-api/util/conversion"
 )
-
-func (src *KubeadmConfig) ConvertTo(dstRaw conversion.Hub) error {
-	dst := dstRaw.(*bootstrapv1.KubeadmConfig)
-	if err := Convert_v1beta1_KubeadmConfig_To_v1beta2_KubeadmConfig(src, dst, nil); err != nil {
-		return err
-	}
-
-	// Manually restore data.
-	restored := &bootstrapv1.KubeadmConfig{}
-	ok, err := utilconversion.UnmarshalData(src, restored)
-	if err != nil {
-		return err
-	}
-
-	// Recover intent for bool values converted to *bool.
-	initialization := bootstrapv1.KubeadmConfigInitializationStatus{}
-	restoredBootstrapDataSecretCreated := restored.Status.Initialization.DataSecretCreated
-	clusterv1.Convert_bool_To_Pointer_bool(src.Status.Ready, ok, restoredBootstrapDataSecretCreated, &initialization.DataSecretCreated)
-	if !reflect.DeepEqual(initialization, bootstrapv1.KubeadmConfigInitializationStatus{}) {
-		dst.Status.Initialization = initialization
-	}
-	if err := RestoreBoolIntentKubeadmConfigSpec(&src.Spec, &dst.Spec, ok, &restored.Spec); err != nil {
-		return err
-	}
-
-	// Recover other values
-	if ok {
-		RestoreKubeadmConfigSpec(&restored.Spec, &dst.Spec)
-	}
-
-	// Override restored data with timeouts values already existing in v1beta1 but in other structs.
-	src.Spec.ConvertTo(&dst.Spec)
-	return nil
-}
-
-func RestoreKubeadmConfigSpec(restored *bootstrapv1.KubeadmConfigSpec, dst *bootstrapv1.KubeadmConfigSpec) {
-	// Restore fields added in v1beta2
-	// Note: Because timeout fields partially exist already in v1beta1 we are using the conversion annotation
-	// instead of backporting the entire timeout fields to v1beta1 and then having some duplicate timeout fields.
-	if restored.InitConfiguration.IsDefined() && !reflect.DeepEqual(restored.InitConfiguration.Timeouts, bootstrapv1.Timeouts{}) {
-		dst.InitConfiguration.Timeouts = restored.InitConfiguration.Timeouts
-	}
-	if restored.JoinConfiguration.IsDefined() && !reflect.DeepEqual(restored.JoinConfiguration.Timeouts, bootstrapv1.Timeouts{}) {
-		dst.JoinConfiguration.Timeouts = restored.JoinConfiguration.Timeouts
-	}
-}
-
-func RestoreBoolIntentKubeadmConfigSpec(src *KubeadmConfigSpec, dst *bootstrapv1.KubeadmConfigSpec, hasRestored bool, restored *bootstrapv1.KubeadmConfigSpec) error {
-	if dst.JoinConfiguration.Discovery.BootstrapToken.IsDefined() {
-		restoredUnsafeSkipCAVerification := restored.JoinConfiguration.Discovery.BootstrapToken.UnsafeSkipCAVerification
-		clusterv1.Convert_bool_To_Pointer_bool(src.JoinConfiguration.Discovery.BootstrapToken.UnsafeSkipCAVerification, hasRestored, restoredUnsafeSkipCAVerification, &dst.JoinConfiguration.Discovery.BootstrapToken.UnsafeSkipCAVerification)
-	}
-
-	if dst.JoinConfiguration.Discovery.File.IsDefined() && dst.JoinConfiguration.Discovery.File.KubeConfig.IsDefined() {
-		if dst.JoinConfiguration.Discovery.File.KubeConfig.Cluster.IsDefined() {
-			restoredInsecureSkipTLSVerify := restored.JoinConfiguration.Discovery.File.KubeConfig.Cluster.InsecureSkipTLSVerify
-			clusterv1.Convert_bool_To_Pointer_bool(src.JoinConfiguration.Discovery.File.KubeConfig.Cluster.InsecureSkipTLSVerify, hasRestored, restoredInsecureSkipTLSVerify, &dst.JoinConfiguration.Discovery.File.KubeConfig.Cluster.InsecureSkipTLSVerify)
-		}
-		if dst.JoinConfiguration.Discovery.File.KubeConfig.User.Exec.IsDefined() {
-			restoredExecProvideClusterInfo := restored.JoinConfiguration.Discovery.File.KubeConfig.User.Exec.ProvideClusterInfo
-			clusterv1.Convert_bool_To_Pointer_bool(src.JoinConfiguration.Discovery.File.KubeConfig.User.Exec.ProvideClusterInfo, hasRestored, restoredExecProvideClusterInfo, &dst.JoinConfiguration.Discovery.File.KubeConfig.User.Exec.ProvideClusterInfo)
-		}
-	}
-
-	for i, volume := range dst.ClusterConfiguration.APIServer.ExtraVolumes {
-		var srcVolume *HostPathMount
-		if src.ClusterConfiguration != nil {
-			for _, v := range src.ClusterConfiguration.APIServer.ExtraVolumes {
-				if v.HostPath == volume.HostPath {
-					srcVolume = &v
-					break
-				}
-			}
-		}
-		if srcVolume == nil {
-			return fmt.Errorf("apiServer extraVolume with hostPath %q not found in source data", volume.HostPath)
-		}
-		var restoredVolumeReadOnly *bool
-		for _, v := range restored.ClusterConfiguration.APIServer.ExtraVolumes {
-			if v.HostPath == volume.HostPath {
-				restoredVolumeReadOnly = v.ReadOnly
-				break
-			}
-		}
-		clusterv1.Convert_bool_To_Pointer_bool(srcVolume.ReadOnly, hasRestored, restoredVolumeReadOnly, &volume.ReadOnly)
-		dst.ClusterConfiguration.APIServer.ExtraVolumes[i] = volume
-	}
-	for i, volume := range dst.ClusterConfiguration.ControllerManager.ExtraVolumes {
-		var srcVolume *HostPathMount
-		if src.ClusterConfiguration != nil {
-			for _, v := range src.ClusterConfiguration.ControllerManager.ExtraVolumes {
-				if v.HostPath == volume.HostPath {
-					srcVolume = &v
-					break
-				}
-			}
-		}
-		if srcVolume == nil {
-			return fmt.Errorf("controllerManager extraVolume with hostPath %q not found in source data", volume.HostPath)
-		}
-		var restoredVolumeReadOnly *bool
-		for _, v := range restored.ClusterConfiguration.ControllerManager.ExtraVolumes {
-			if v.HostPath == volume.HostPath {
-				restoredVolumeReadOnly = v.ReadOnly
-				break
-			}
-		}
-		clusterv1.Convert_bool_To_Pointer_bool(srcVolume.ReadOnly, hasRestored, restoredVolumeReadOnly, &volume.ReadOnly)
-		dst.ClusterConfiguration.ControllerManager.ExtraVolumes[i] = volume
-	}
-	for i, volume := range dst.ClusterConfiguration.Scheduler.ExtraVolumes {
-		var srcVolume *HostPathMount
-		if src.ClusterConfiguration != nil {
-			for _, v := range src.ClusterConfiguration.Scheduler.ExtraVolumes {
-				if v.HostPath == volume.HostPath {
-					srcVolume = &v
-					break
-				}
-			}
-		}
-		if srcVolume == nil {
-			return fmt.Errorf("scheduler extraVolume with hostPath %q not found in source data", volume.HostPath)
-		}
-		var restoredVolumeReadOnly *bool
-		for _, v := range restored.ClusterConfiguration.Scheduler.ExtraVolumes {
-			if v.HostPath == volume.HostPath {
-				restoredVolumeReadOnly = v.ReadOnly
-				break
-			}
-		}
-		clusterv1.Convert_bool_To_Pointer_bool(srcVolume.ReadOnly, hasRestored, restoredVolumeReadOnly, &volume.ReadOnly)
-		dst.ClusterConfiguration.Scheduler.ExtraVolumes[i] = volume
-	}
-
-	for i, file := range dst.Files {
-		var srcFile *File
-		for _, f := range src.Files {
-			if f.Path == file.Path {
-				srcFile = &f
-				break
-			}
-		}
-		if srcFile == nil {
-			return fmt.Errorf("file with path %q not found in source data", file.Path)
-		}
-		var restoredFileAppend *bool
-		for _, f := range restored.Files {
-			if f.Path == file.Path {
-				restoredFileAppend = f.Append
-				break
-			}
-		}
-		clusterv1.Convert_bool_To_Pointer_bool(srcFile.Append, hasRestored, restoredFileAppend, &file.Append)
-		dst.Files[i] = file
-	}
-
-	if dst.Ignition.IsDefined() && dst.Ignition.ContainerLinuxConfig.IsDefined() {
-		restoredIgnitionStrict := restored.Ignition.ContainerLinuxConfig.Strict
-		clusterv1.Convert_bool_To_Pointer_bool(src.Ignition.ContainerLinuxConfig.Strict, hasRestored, restoredIgnitionStrict, &dst.Ignition.ContainerLinuxConfig.Strict)
-	}
-	return nil
-}
-
-func (src *KubeadmConfigSpec) ConvertTo(dst *bootstrapv1.KubeadmConfigSpec) {
-	// Override with timeouts values already existing in v1beta1.
-	var initControlPlaneComponentHealthCheckSeconds *int32
-	if src.ClusterConfiguration != nil && src.ClusterConfiguration.APIServer.TimeoutForControlPlane != nil {
-		dst.InitConfiguration.Timeouts.ControlPlaneComponentHealthCheckSeconds = clusterv1.ConvertToSeconds(src.ClusterConfiguration.APIServer.TimeoutForControlPlane)
-		initControlPlaneComponentHealthCheckSeconds = dst.InitConfiguration.Timeouts.ControlPlaneComponentHealthCheckSeconds
-	}
-	if (src.JoinConfiguration != nil && src.JoinConfiguration.Discovery.Timeout != nil) || initControlPlaneComponentHealthCheckSeconds != nil {
-		dst.JoinConfiguration.Timeouts.ControlPlaneComponentHealthCheckSeconds = initControlPlaneComponentHealthCheckSeconds
-		if src.JoinConfiguration != nil && src.JoinConfiguration.Discovery.Timeout != nil {
-			dst.JoinConfiguration.Timeouts.TLSBootstrapSeconds = clusterv1.ConvertToSeconds(src.JoinConfiguration.Discovery.Timeout)
-		}
-	}
-}
-
-func (dst *KubeadmConfig) ConvertFrom(srcRaw conversion.Hub) error {
-	src := srcRaw.(*bootstrapv1.KubeadmConfig)
-	if err := Convert_v1beta2_KubeadmConfig_To_v1beta1_KubeadmConfig(src, dst, nil); err != nil {
-		return err
-	}
-
-	// Convert timeouts moved from one struct to another.
-	dst.Spec.ConvertFrom(&src.Spec)
-
-	dropEmptyStringsKubeadmConfigSpec(&dst.Spec)
-	dropEmptyStringsKubeadmConfigStatus(&dst.Status)
-
-	// Preserve Hub data on down-conversion except for metadata.
-	return utilconversion.MarshalDataUnsafeNoCopy(src, dst)
-}
-
-func (dst *KubeadmConfigSpec) ConvertFrom(src *bootstrapv1.KubeadmConfigSpec) {
-	// Convert timeouts moved from one struct to another.
-	if src.InitConfiguration.Timeouts.ControlPlaneComponentHealthCheckSeconds != nil {
-		if dst.ClusterConfiguration == nil {
-			dst.ClusterConfiguration = &ClusterConfiguration{}
-		}
-		dst.ClusterConfiguration.APIServer.TimeoutForControlPlane = clusterv1.ConvertFromSeconds(src.InitConfiguration.Timeouts.ControlPlaneComponentHealthCheckSeconds)
-	}
-	if reflect.DeepEqual(dst.InitConfiguration, &InitConfiguration{}) {
-		dst.InitConfiguration = nil
-	}
-	if src.JoinConfiguration.Timeouts.TLSBootstrapSeconds != nil {
-		if dst.JoinConfiguration == nil {
-			dst.JoinConfiguration = &JoinConfiguration{}
-		}
-		dst.JoinConfiguration.Discovery.Timeout = clusterv1.ConvertFromSeconds(src.JoinConfiguration.Timeouts.TLSBootstrapSeconds)
-	}
-	if reflect.DeepEqual(dst.JoinConfiguration, &JoinConfiguration{}) {
-		dst.JoinConfiguration = nil
-	}
-}
-
-func (src *KubeadmConfigTemplate) ConvertTo(dstRaw conversion.Hub) error {
-	dst := dstRaw.(*bootstrapv1.KubeadmConfigTemplate)
-	if err := Convert_v1beta1_KubeadmConfigTemplate_To_v1beta2_KubeadmConfigTemplate(src, dst, nil); err != nil {
-		return err
-	}
-
-	// Manually restore data.
-	restored := &bootstrapv1.KubeadmConfigTemplate{}
-	ok, err := utilconversion.UnmarshalData(src, restored)
-	if err != nil {
-		return err
-	}
-	// Recover intent for bool values converted to *bool.
-	if err := RestoreBoolIntentKubeadmConfigSpec(&src.Spec.Template.Spec, &dst.Spec.Template.Spec, ok, &restored.Spec.Template.Spec); err != nil {
-		return err
-	}
-
-	// Recover other values
-	if ok {
-		RestoreKubeadmConfigSpec(&restored.Spec.Template.Spec, &dst.Spec.Template.Spec)
-	}
-
-	// Override restored data with timeouts values already existing in v1beta1 but in other structs.
-	src.Spec.Template.Spec.ConvertTo(&dst.Spec.Template.Spec)
-	return nil
-}
-
-func (dst *KubeadmConfigTemplate) ConvertFrom(srcRaw conversion.Hub) error {
-	src := srcRaw.(*bootstrapv1.KubeadmConfigTemplate)
-	if err := Convert_v1beta2_KubeadmConfigTemplate_To_v1beta1_KubeadmConfigTemplate(src, dst, nil); err != nil {
-		return err
-	}
-
-	// Convert timeouts moved from one struct to another.
-	dst.Spec.Template.Spec.ConvertFrom(&src.Spec.Template.Spec)
-
-	dropEmptyStringsKubeadmConfigSpec(&dst.Spec.Template.Spec)
-
-	// Preserve Hub data on down-conversion except for metadata.
-	return utilconversion.MarshalDataUnsafeNoCopy(src, dst)
-}
 
 func Convert_v1beta2_InitConfiguration_To_v1beta1_InitConfiguration(in *bootstrapv1.InitConfiguration, out *InitConfiguration, s apimachineryconversion.Scope) error {
 	// Timeouts requires conversion at an upper level
@@ -951,39 +691,4 @@ func Convert_v1beta2_File_To_v1beta1_File(in *bootstrapv1.File, out *File, s api
 		}
 	}
 	return nil
-}
-
-func dropEmptyStringsKubeadmConfigSpec(dst *KubeadmConfigSpec) {
-	for i, u := range dst.Users {
-		dropEmptyString(&u.Gecos)
-		dropEmptyString(&u.Groups)
-		dropEmptyString(&u.HomeDir)
-		dropEmptyString(&u.Shell)
-		dropEmptyString(&u.Passwd)
-		dropEmptyString(&u.PrimaryGroup)
-		dropEmptyString(&u.Sudo)
-		dst.Users[i] = u
-	}
-
-	if dst.DiskSetup != nil {
-		for i, p := range dst.DiskSetup.Partitions {
-			dropEmptyString(&p.TableType)
-			dst.DiskSetup.Partitions[i] = p
-		}
-		for i, f := range dst.DiskSetup.Filesystems {
-			dropEmptyString(&f.Partition)
-			dropEmptyString(&f.ReplaceFS)
-			dst.DiskSetup.Filesystems[i] = f
-		}
-	}
-}
-
-func dropEmptyStringsKubeadmConfigStatus(dst *KubeadmConfigStatus) {
-	dropEmptyString(&dst.DataSecretName)
-}
-
-func dropEmptyString(s **string) {
-	if *s != nil && **s == "" {
-		*s = nil
-	}
 }
