+++ controlplane/kubeadm/webhooks/conversion/common.go
@@ -1,0 +1,34 @@
+/*
+Copyright 2026 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package conversion
+
+import (
+	"context"
+	"errors"
+
+	"k8s.io/apimachinery/pkg/runtime/schema"
+)
+
+var apiVersionGetter = func(_ context.Context, _ schema.GroupKind) (string, error) {
+	return "", errors.New("apiVersionGetter not set")
+}
+
+// SetAPIVersionGetter sets an APIVersionGetter that is required during conversion to retrieve
+// the APIVersion when converting from v1beta2 to v1beta1.
+func SetAPIVersionGetter(f func(ctx context.Context, gk schema.GroupKind) (string, error)) {
+	apiVersionGetter = f
+}
