+++ controlplane/kubeadm/webhooks/conversion/doc.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2025 The Kubernetes Authors.
+Copyright 2026 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,7 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1beta2
-
-func (*KubeadmControlPlane) Hub()         {}
-func (*KubeadmControlPlaneTemplate) Hub() {}
+// Package conversion provides conversion code for all APIs of the kubeadm controlplane CAPI provider.
+package conversion
